till-boot-config version 17
New configuration scheme with "drivers" and "roles"

Currently there are two classes of driver: "generic" and "till-boot".
"generic" is not useful yet — it will become so once till-boot-config
can also generate network and dhcp configuration files. "till-boot"
encapsulates the previous behaviour.

It's possible to create multiple instances of driver classes, with
different configurations. Again, not particularly useful yet but may
become so in the future.

The "roles" configuration is useful immediately: think of it as named
sets of defaults. So at EMF, for example, we could have a role for
"till" and a role for "displayboard" with appropriate defaults for each.

Multiple roles can be specified for a till, and the defaults from each
role will be applied in order.