Create quicktill configuration file
If option "quicktill-config" is present, it is written to
~till/.config/quicktill.json

This is intended to be used for the configurl and database options, but
could have other uses.

The older "configurl", "dbname", etc. options are still supported for
now, but a warning is output at till startup time if quicktill-config
isn't present in the configuration.